code linting and changed tag number-strings
