+++ src/Components/Collections/Collection.js
@@ -6,11 +6,9 @@ import React from 'react';
 import {ApplicationStore, ThingsdbStore} from '../../Stores';
 import {CollectionConfig} from './Config';
 import {findItem, HarmonicCardHeader, isObjectEmpty, TitlePage} from '../Util';
-import {Procedures, Timers} from './ProceduresTimers';
+import {Procedures, Timers} from '../ProceduresAndTimers';
 import CollectionTree from './Tree';
 import CollectionEnumsTypes from './EnumsTypes';
-import {ProcedureActions, TimerActions} from '../../Stores';
-import {ProceduresTAG, TimersTAG} from '../../constants';
 
 
 const withStores = withVlow([{
@@ -31,14 +29,6 @@ const Collection = ({match, collections}) => {
     const classes = useStyles();
     const selectedCollection = findItem(match.index, collections);
 
-    const handleRefreshProcedures = () => {
-        ProcedureActions.getProcedures(`@collection:${selectedCollection.name}`, ProceduresTAG);
-    };
-
-    const handleRefreshTimers = () => {
-        TimerActions.getTimers(`@collection:${selectedCollection.name}`, TimersTAG);
-    };
-
     return (
         isObjectEmpty(selectedCollection) ? null : (
             <TitlePage
@@ -58,14 +48,18 @@ const Collection = ({match, collections}) => {
                         </Grid>
                         <Grid container item md={5} xs={12}>
                             <Grid className={classes.spacing} item xs={12}>
-                                <HarmonicCardHeader title="PROCEDURES" onRefresh={handleRefreshProcedures} unmountOnExit>
-                                    <Procedures scope={`@collection:${selectedCollection.name}`} />
-                                </HarmonicCardHeader >
+                                <Procedures
+                                    buttonsView={{add: true, edit: true, run: true, view: true}}
+                                    dialogsView={{add: true, edit: true, run: true, view: true}}
+                                    scope={`@collection:${selectedCollection.name}`}
+                                />
                             </Grid>
                             <Grid className={classes.spacing} item xs={12}>
-                                <HarmonicCardHeader title="TIMERS" onRefresh={handleRefreshTimers} unmountOnExit>
-                                    <Timers scope={`@collection:${selectedCollection.name}`} />
-                                </HarmonicCardHeader >
+                                <Timers
+                                    buttonsView={{add: true, edit: true, run: true, view: true}}
+                                    dialogsView={{add: true, edit: true, run: true, view: true}}
+                                    scope={`@collection:${selectedCollection.name}`}
+                                />
                             </Grid>
                             <CollectionEnumsTypes scope={`@collection:${selectedCollection.name}`} />
                         </Grid>
