--- src/Components/Collections/ProceduresTimers/index.js
@@ -1,7 +1,0 @@
-import Procedures from './Procedures';
-import Timers from './Timers';
-
-export {
-    Procedures,
-    Timers
-};
