+++ src/Components/Editor/EditorSideContent/EnumsTypes.js
@@ -1,0 +1,157 @@
+import {withVlow} from 'vlow';
+import Grid from '@material-ui/core/Grid';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import {EnumActions, TypeActions, EnumStore, TypeStore} from '../../../Stores';
+import {HarmonicCardHeader, WarnPopover} from '../../Util';
+import {EnumTypeChips} from '../../Collections/CollectionsUtils/TypesEnumsUtils';
+import {EnumsTAG, TypesTAG} from '../../../constants';
+
+const withStores = withVlow([{
+    store: EnumStore,
+    keys: ['enums']
+}, {
+    store: TypeStore,
+    keys: ['customTypes']
+}]);
+
+const EnumsTypes = ({customTypes, enums, onSetQueryInput, scope}) => {
+    const [viewEnum, setViewEnum] = React.useState({
+        open: false,
+        name: '',
+        expand: false,
+    });
+    const [viewType, setViewType] = React.useState({
+        open: false,
+        name: '',
+        expand: false,
+    });
+
+    const [anchorEl, setAnchorEl] = React.useState(null);
+    const [warnDescription, setWarnDescription] = React.useState('');
+
+    const handleRefreshEnums = React.useCallback(() => EnumActions.getEnums(scope, EnumsTAG), [scope]);
+    const handleRefreshTypes = React.useCallback(() => TypeActions.getTypes(scope, TypesTAG), [scope]);
+
+    React.useEffect(() => {
+        handleRefreshTypes();
+        handleRefreshEnums();
+    }, [handleRefreshEnums, handleRefreshTypes,]);
+
+    const makeTypeInstanceInit = (n, customTypeNames, customTypes, circularRefFlag, target) => {
+        if (customTypeNames.includes(n)) {
+            if (circularRefFlag[n]) {
+                setAnchorEl(target);
+                setWarnDescription('Circular reference detected');
+                return '';
+            } else {
+                circularRefFlag[n] = true;
+                return `${n}{${customTypes.find(i=>i.name==n).fields.map(c =>`${c[0]}: ${makeTypeInstanceInit(c[1], customTypeNames, customTypes, {...circularRefFlag}, target)}`)}}`;
+            }
+        }
+        return `<${n}>`;
+    };
+    const makeEnumInstanceInit = (n)  => {
+        return `${n}{...}`;
+    };
+
+    const handleChange = (a) => (n, c) => {
+        switch(a){
+        case 'view':
+            if (c=='type') {
+                setViewEnum({...viewEnum, open: false, name: ''});
+                setViewType({open: true, expand: true, name: n});
+            } else {
+                setViewType({...viewType, open: false, name: ''});
+                setViewEnum({open: true, expand: true, name: n});
+            }
+            break;
+        case 'add':
+            if (c=='type') {
+                onSetQueryInput('set_type("...", {...})');
+            } else {
+                onSetQueryInput('set_enum("...", {...})');
+            }
+            break;
+        }
+    };
+
+    const handleClose = (_a, c) => {
+        if (c=='type') {
+            setViewType({...viewType, open: false, name: ''});
+        } else {
+            setViewEnum({...viewEnum, open: false, name: ''});
+        }
+    };
+
+    const handleExpand = (ky) => (check) => {
+        switch(ky){
+        case 'type':
+            setViewType({...viewType, expand: check});
+            break;
+        case 'enum':
+            setViewEnum({...viewEnum, expand: check});
+            break;
+        }
+    };
+
+    const handleCloseWarn = () => {
+        setAnchorEl(null);
+    };
+
+    return (
+        <React.Fragment>
+            <WarnPopover anchorEl={anchorEl} onClose={handleCloseWarn} description={warnDescription} />
+            <Grid item xs={12}>
+                <HarmonicCardHeader expand={viewType.expand} onExpand={handleExpand('type')} title="TYPES" onRefresh={handleRefreshTypes} unmountOnExit>
+                    <EnumTypeChips
+                        buttonsView={{add: false, edit: false, run: true, view: true}}
+                        categoryInit="type"
+                        items={customTypes[scope]}
+                        onChange={handleChange}
+                        onClose={handleClose}
+                        onDelete={TypeActions.deleteType}
+                        onInfo={TypeActions.getTypes}
+                        onMakeInstanceInit={makeTypeInstanceInit}
+                        onSetQueryInput={onSetQueryInput}
+                        scope={scope}
+                        tag={TypesTAG}
+                        view={{view: viewType.open, name: viewType.name}}
+                    />
+                </HarmonicCardHeader>
+            </Grid>
+            <Grid item xs={12}>
+                <HarmonicCardHeader expand={viewEnum.expand} onExpand={handleExpand('enums')} title="ENUMS" onRefresh={handleRefreshEnums} unmountOnExit>
+                    <EnumTypeChips
+                        buttonsView={{add: false, edit: false, run: true, view: true}}
+                        categoryInit="enum"
+                        items={enums[scope]}
+                        onChange={handleChange}
+                        onClose={handleClose}
+                        onDelete={EnumActions.deleteEnum}
+                        onInfo={EnumActions.getEnums}
+                        onMakeInstanceInit={makeEnumInstanceInit}
+                        onSetQueryInput={onSetQueryInput}
+                        scope={scope}
+                        tag={EnumsTAG}
+                        view={{view: viewEnum.open, name: viewEnum.name}}
+                    />
+                </HarmonicCardHeader>
+            </Grid>
+        </React.Fragment>
+
+    );
+};
+
+EnumsTypes.propTypes = {
+    onSetQueryInput: PropTypes.func.isRequired,
+    scope: PropTypes.string.isRequired,
+
+    /* types properties */
+    customTypes: TypeStore.types.customTypes.isRequired,
+    /* enums properties */
+    enums: EnumStore.types.enums.isRequired,
+};
+
+export default withStores(EnumsTypes);
