+++ src/Components/ProceduresAndTimers/Utils/index.js
@@ -1,0 +1,7 @@
+import Card from './Card';
+import Page from './Page';
+
+export {
+    Card,
+    Page
+};
