+++ src/Components/ProceduresAndTimers/index.js
@@ -1,6 +1,17 @@
 import {AddProcedureDialog, ProcedureDialogs} from './ProcedureDialogs';
 import {AddTimerDialog, TimerDialogs} from './TimerDialogs';
 import Procedure from './Procedure';
+import Procedures from './Procedures';
 import Timer from './Timer';
+import Timers from './Timers';
 
-export {AddProcedureDialog, AddTimerDialog, Procedure, ProcedureDialogs, Timer, TimerDialogs};
+export {
+    AddProcedureDialog,
+    AddTimerDialog,
+    Procedure,
+    ProcedureDialogs,
+    Procedures,
+    Timer,
+    TimerDialogs,
+    Timers,
+};
