+++ src/constants.js
@@ -1,44 +1,44 @@
-export const AuthTAG = '0';
-export const LoginTAG = '1';
-export const AddCollectionTAG = '2';
-export const QuotasTAG = '3';
-export const RemoveCollectionTAG = '4';
-export const RenameCollectionTAG = '5';
-export const ProceduresTAG = '6';
-export const TimersTAG = '7';
-export const RemoveThingTAG = '8';
-export const ThingActionsDialogTAG = '9';
-export const EnumsTAG = '10';
-export const TypesTAG = '11';
-export const AddDialogTAG = '12';
-export const EditDialogTAG = '13';
-export const EditorTAG = '14';
-export const AddBackupTAG = '15';
-export const AddNodeTAG = '16';
-export const LoglevelTAG = '17';
-export const RemoveNodeTAG = '18';
-export const ShutdownTAG = '19';
-export const RemoveProcedureTAG = '20';
-export const RunProcedureTAG = '21';
-export const AddProcedureDialogTAG = '22';
-export const EditProcedureDialogTAG = '23';
-export const RunProcedureDialogTAG = '24';
-export const AddUserTAG = '25';
-export const PasswordTAG = '26';
-export const RemoveUserTAG = '27';
-export const RenameUserTAG = '28';
-export const UserAccessTAG = '29';
-export const AddTokenTAG = '30';
-export const TokensTAG = '31';
-export const ChipsCardTAG = '32';
-export const WatcherTAG = '33';
-export const RemoveBackupTAG = '34';
-export const RestoreNodeTAG = '35';
-export const ExportCollectionTAG = '36';
-export const AddModuleTAG = '37';
-export const RemoveModuleTAG = '38';
-export const ModuleInfoTAG = '39';
-export const AddTimerDialogTAG = '40';
-export const EditTimerDialogTAG = '41';
-export const RunTimerDialogTAG = '42';
-export const ViewTimerDialogTAG = '43';
+export const AuthTAG = 'AuthTAG';
+export const LoginTAG = 'LoginTAG';
+export const AddCollectionTAG = 'AddCollectionTAG';
+export const QuotasTAG = 'QuotasTAG';
+export const RemoveCollectionTAG = 'RemoveCollectionTAG';
+export const RenameCollectionTAG = 'RenameCollectionTAG';
+export const ProceduresTAG = 'ProceduresTAG';
+export const TimersTAG = 'TimersTAG';
+export const RemoveThingTAG = 'RemoveThingTAG';
+export const ThingActionsDialogTAG = 'ThingActionsDialogTAG';
+export const EnumsTAG = 'EnumsTAG';
+export const TypesTAG = 'TypesTAG';
+export const AddDialogTAG = 'AddDialogTAG';
+export const EditDialogTAG = 'EditDialogTAG';
+export const EditorTAG = 'EditorTAG';
+export const AddBackupTAG = 'AddBackupTAG';
+export const AddNodeTAG = 'AddNodeTAG';
+export const LoglevelTAG = 'LoglevelTAG';
+export const RemoveNodeTAG = 'RemoveNodeTAG';
+export const ShutdownTAG = 'ShutdownTAG';
+export const RemoveProcedureTAG = 'RemoveProcedureTAG';
+export const RunProcedureTAG = 'RunProcedureTAG';
+export const AddProcedureDialogTAG = 'AddProcedureDialogTAG';
+export const EditProcedureDialogTAG = 'EditProcedureDialogTAG';
+export const RunProcedureDialogTAG = 'RunProcedureDialogTAG';
+export const AddUserTAG = 'AddUserTAG';
+export const PasswordTAG = 'PasswordTAG';
+export const RemoveUserTAG = 'RemoveUserTAG';
+export const RenameUserTAG = 'RenameUserTAG';
+export const UserAccessTAG = 'UserAccessTAG';
+export const AddTokenTAG = 'AddTokenTAG';
+export const TokensTAG = 'TokensTAG';
+export const ChipsCardTAG = 'ChipsCardTAG';
+export const WatcherTAG = 'WatcherTAG';
+export const RemoveBackupTAG = 'RemoveBackupTAG';
+export const RestoreNodeTAG = 'RestoreNodeTAG';
+export const ExportCollectionTAG = 'ExportCollectionTAG';
+export const AddModuleTAG = 'AddModuleTAG';
+export const RemoveModuleTAG = 'RemoveModuleTAG';
+export const ModuleInfoTAG = 'ModuleInfoTAG';
+export const AddTimerDialogTAG = 'AddTimerDialogTAG';
+export const EditTimerDialogTAG = 'EditTimerDialogTAG';
+export const RunTimerDialogTAG = 'RunTimerDialogTAG';
+export const ViewTimerDialogTAG = 'ViewTimerDialogTAG';
